CB-10-DIRETOR ENTITY

+++ src/main/java/com/example/compassobank/dto/ContaDTO.java
@@ -2,6 +2,8 @@ package com.example.compassobank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContaDTO {
 
@@ -9,10 +11,10 @@ public class ContaDTO {
 
     private String agencia;
 
-    private double saldo;
+    private BigDecimal saldo;
 
     private String senha;
 
-    private double juros;
+    private BigDecimal juros;
 
 }
