+++ src/main/java/com/example/compassobank/dto/ContaPessoalDTO.java
@@ -2,6 +2,8 @@ package com.example.compassobank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContaPessoalDTO {
 
@@ -9,17 +11,17 @@ public class ContaPessoalDTO {
 
     private String agencia;
 
-    private double saldo;
+    private BigDecimal saldo;
 
     private String senha;
 
-    private double juros;
+    private BigDecimal juros;
 
-    private double limite;
+    private BigDecimal limite;
 
     private int cheques;
 
-    private double credito;
+    private BigDecimal credito;
 
-    private double saldoAplicacao;
+    private BigDecimal saldoAplicacao;
 }
